[STRUCTURAL] Move Method: strategy contribution to a shared seam runtime
strategySectionsRecovering was identical in both cores, and the loop
around it differed only in its sink: internal/agent feeds a PromptBuilder,
the SDK returns the slice. The iteration, the blank-content skip and the
per-strategy recover boundary are the parts worth having once.

agentsdk.ContributeSections now owns all three and returns the sections;
the SDK's method is a one-line delegate, and internal's keeps its
PromptBuilder sink as a loop over the shared result.

Same reasoning as the background dispatch: a recover boundary kept in two
copies is a boundary that gets fixed in one.

No behavior change — blank-skip and ordering are preserved, and the four
built-in strategies' tests cover the internal path unchanged.
