+++ internal/agent/context_strategy.go
@@ -74,25 +74,12 @@ func (a *Agent) staticSectionsRecovering(source agentsdk.StaticPromptSource) (se
 // running on the turn goroutine, where an unrecovered panic would abort
 // the user turn and starve sibling strategies.
 func (a *Agent) contributeStrategySections(ctx context.Context, pb *PromptBuilder, info agentsdk.PromptContext) {
-	for _, strategy := range a.contextStrategies {
-		for _, section := range a.strategySectionsRecovering(ctx, strategy, info) {
-			if strings.TrimSpace(section.Content) == "" {
-				continue
-			}
-			pb.AddDynamicSection_UNCACHED(section.Title, section.Content, section.Reason)
-		}
+	// The iteration, blank-skip and per-strategy recover boundary are shared
+	// with the portable loop; only the sink differs — this core feeds a
+	// PromptBuilder, the SDK returns the slice to its caller.
+	for _, section := range agentsdk.ContributeSections(ctx, a.contextStrategies, info, a.logger) {
+		pb.AddDynamicSection_UNCACHED(section.Title, section.Content, section.Reason)
 	}
-}
-
-// strategySectionsRecovering invokes one strategy behind a recover
-// boundary; on panic the strategy contributes nothing this turn.
-func (a *Agent) strategySectionsRecovering(ctx context.Context, strategy agentsdk.ContextStrategy, info agentsdk.PromptContext) (sections []agentsdk.PromptSection) {
-	defer func() {
-		if r := recover(); r != nil {
-			a.logger.Warn("context strategy ContributePromptSections panicked: %v", r)
-		}
-	}()
-	return strategy.ContributePromptSections(ctx, info)
 }
 
 // builtinContextStrategies returns the agent's built-in prompt
